Refs #599. Fixed compilation in Windows.

+++ src/transports/dds/components/ProxyProcedureEndpoint.cpp
@@ -383,7 +383,7 @@ ReturnMessage ProxyProcedureEndpoint::send(void *request, void *reply)
     DDS::WaitSet *waitSet = NULL;
     DDS::ReturnCode_t retCode;
     boost::posix_time::time_duration tTimeout = boost::posix_time::milliseconds(m_transport.getTimeout());
-    unsigned int numSec = 0;
+    int64_t numSec = 0;
     char value[50];
     char *auxPointerToRequest = NULL, *seqAuxPointer = NULL;
     char **auxPointerToRemoteServiceName = NULL;
@@ -544,7 +544,7 @@ ReturnMessage ProxyProcedureEndpoint::send_async(void *request, DDSAsyncTask *ta
     ReturnMessage returnedValue = CLIENT_INTERNAL_ERROR;
     DDS::WaitSet *waitSet = NULL;
     DDS::ReturnCode_t retCode;
-    unsigned int numSec = 0;
+    int64_t numSec = 0;
     char value[50];
     char *auxPointerToRequest = NULL, *seqAuxPointer = NULL;
     char **auxPointerToRemoteServiceName = NULL;
